requirements.txt
asgiref==3.7.2
blinker==1.4
certifi==2023.7.22

cryptography==3.4.8

distlib==0.3.7
distro==1.7.0

Django==4.2.4
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
filelock==3.12.2
httplib2==0.20.2
importlib-metadata==4.6.4
inflection==0.5.1
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Markdown==3.4.4
more-itertools==8.10.0

oauthlib==3.2.0
packaging==24.0
pipenv==2023.8.23
platformdirs==3.10.0

PyJWT==2.3.0
pyparsing==2.4.7

pytz==2022.1

SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.2

typing_extensions==4.7.1



uritemplate==4.1.1
virtualenv==20.24.3
wadllib==1.3.6
zipp==1.0.0
